Add files via upload
he application is designed to mimic the Amazon app. The app is made using Swift programming language, Storyboard and UIKit. There is a search bar and a link bar on the top. The link bar scrolls horizontally, to show more links. The links when tapped segues to a mock page, with webview. The main/home page shows the webview which loads amazon.com. The tab bar consists of two tab items - home and cart. Home tab loads amazon home page while cart tab loads the amazon cart page. By default home tab is selected. Animation is used when the webview scrolls up and down, the tab bar and the links to disappear and reappear.